Update
A message when the game is over has been added.
The constructor of Snake class has been modified just for testing purposes.
The Update method of Game class has been modified to close the window once the snake is no longer alive. This is achieved by passing the running variable by reference to the method and change the value to false.